Move `Digestible[UUID]` spec to `DigestibleSpec`
