+++ memeid4s/src/test/scala/memeid4s/UUIDSpec.scala
@@ -16,11 +16,8 @@
 
 package memeid4s
 
-import java.nio.ByteBuffer
-
 import memeid4s.UUID.RichUUID
 import memeid.arbitrary.instances._
-import memeid4s.digest.Digestible
 import org.scalacheck.Gen
 import org.specs2.ScalaCheck
 import org.specs2.mutable.Specification
@@ -154,27 +151,6 @@ class UUIDSpec extends Specification with ScalaCheck {
 
     "have all 128 bits to 0" in {
       (UUID.Nil.msb must be equalTo 0L) and (UUID.Nil.lsb must be equalTo 0L)
-    }
-
-  }
-
-  "Digestible[UUID]" should {
-
-    def fromByteArray(bytes: Array[Byte]): UUID = {
-      val bb = ByteBuffer.wrap(bytes)
-      UUID.from(bb.getLong, bb.getLong)
-    }
-
-    "round-trip" in prop { (msb: Long, lsb: Long) =>
-      val uuid  = UUID.from(msb, lsb)
-      val bytes = Digestible[UUID].toByteArray(uuid)
-      fromByteArray(bytes) must be equalTo uuid
-    }
-
-    "give the same bytes for the same UUID" in {
-      val uuid = UUID.V1.next
-
-      Digestible[UUID].toByteArray(uuid) must be equalTo Digestible[UUID].toByteArray(uuid)
     }
 
   }
